removed test_11 and test_12 warnings

+++ world.cpp
@@ -387,16 +387,16 @@ unsigned int World::count_neighbours(int x, int y, bool toroidal) {
                 if (b==-1) {
                     temp_b=current_state.get_height()+b;
                 } 
-                if (a==get_width()) {
+                if (a==(int)get_width()) {
                     temp_a=0;
                 }
-                if (b==get_height()) {
+                if (b==(int)get_height()) {
                     temp_b=0;
                 }
 
             } else {
                 //fix out of bounds
-                if (((a==-1)) || (b==-1) || (a==get_width()) || (b==get_height()) ) {
+                if (((a==-1)) || (b==-1) || (a==(int)get_width()) || (b==(int)get_height()) ) {
                     temp_a=x;
                     temp_b=y;
                 }
